Moving some things around and renaming things.
