+++ src/mpfao/tdycoremesh.c
@@ -5,7 +5,6 @@
 #include <private/tdymeshutilsimpl.h>
 #include <private/tdymemoryimpl.h>
 #include <private/tdyutils.h>
-#include <private/tdymeshutilsimpl.h>
 #include <private/tdyregionimpl.h>
 #include <private/tdydiscretization.h>
 
@@ -14,339 +13,6 @@
 PetscBool IsClosureWithinBounds(PetscInt closure, PetscInt start,
                                 PetscInt end) {
   return (closure >= start) && (closure < end);
-}
-
-/* -------------------------------------------------------------------------- */
-PetscErrorCode AllocateMemoryForCells(
-  PetscInt    num_cells,
-  TDyCellType cell_type,
-  TDyCell    *cells) {
-
-  PetscFunctionBegin;
-
-  PetscInt num_vertices  = GetNumVerticesForCellType(cell_type);
-  PetscInt num_edges     = GetNumEdgesForCellType(cell_type);
-  PetscInt num_neighbors = GetNumNeighborsForCellType(cell_type);
-  PetscInt num_faces     = GetNumFacesForCellType(cell_type);
-
-  TDySubcellType subcell_type = GetSubcellTypeForCellType(cell_type);
-  PetscInt num_subcells = GetNumSubcellsForSubcellType(subcell_type);
-  num_subcells = num_vertices;
-
-  PetscErrorCode ierr;
-  ierr = TDyAllocate_IntegerArray_1D(&cells->id,num_cells); CHKERRQ(ierr);
-  ierr = TDyAllocate_IntegerArray_1D(&cells->global_id,num_cells); CHKERRQ(ierr);
-  ierr = TDyAllocate_IntegerArray_1D(&cells->natural_id,num_cells); CHKERRQ(ierr);
-
-   cells->is_local = (PetscBool *)malloc(num_cells*sizeof(PetscBool));
-
-  ierr = TDyAllocate_IntegerArray_1D(&cells->num_vertices,num_cells); CHKERRQ(ierr);
-  ierr = TDyAllocate_IntegerArray_1D(&cells->num_edges,num_cells); CHKERRQ(ierr);
-  ierr = TDyAllocate_IntegerArray_1D(&cells->num_faces,num_cells); CHKERRQ(ierr);
-  ierr = TDyAllocate_IntegerArray_1D(&cells->num_neighbors,num_cells); CHKERRQ(ierr);
-  ierr = TDyAllocate_IntegerArray_1D(&cells->num_subcells,num_cells); CHKERRQ(ierr);
-
-  ierr = TDyAllocate_IntegerArray_1D(&cells->vertex_offset  ,num_cells+1); CHKERRQ(ierr);
-  ierr = TDyAllocate_IntegerArray_1D(&cells->edge_offset    ,num_cells+1); CHKERRQ(ierr);
-  ierr = TDyAllocate_IntegerArray_1D(&cells->face_offset    ,num_cells+1); CHKERRQ(ierr);
-  ierr = TDyAllocate_IntegerArray_1D(&cells->neighbor_offset,num_cells+1); CHKERRQ(ierr);
-
-  ierr = TDyAllocate_IntegerArray_1D(&cells->vertex_ids  ,num_cells*num_vertices); CHKERRQ(ierr);
-  ierr = TDyAllocate_IntegerArray_1D(&cells->edge_ids    ,num_cells*num_edges); CHKERRQ(ierr);
-  ierr = TDyAllocate_IntegerArray_1D(&cells->neighbor_ids,num_cells*num_neighbors); CHKERRQ(ierr);
-  ierr = TDyAllocate_IntegerArray_1D(&cells->face_ids    ,num_cells*num_faces); CHKERRQ(ierr);
-
-  ierr = TDyAllocate_TDyCoordinate_1D(num_cells,&cells->centroid); CHKERRQ(ierr);
-
-  ierr = TDyAllocate_RealArray_1D(&cells->volume,num_cells); CHKERRQ(ierr);
-
-  for (PetscInt icell=0; icell<num_cells; icell++) {
-    cells->id[icell]            = icell;
-    cells->num_vertices[icell]  = num_vertices;
-    cells->num_edges[icell]     = num_edges;
-    cells->num_faces[icell]     = num_faces;
-    cells->num_neighbors[icell] = num_neighbors;
-    cells->num_subcells[icell]  = num_subcells;
-  }
-
-  for (PetscInt icell=0; icell<=num_cells; icell++) {
-    cells->vertex_offset[icell]   = icell*num_vertices;
-    cells->edge_offset[icell]     = icell*num_edges;
-    cells->face_offset[icell]     = icell*num_faces;
-    cells->neighbor_offset[icell] = icell*num_neighbors;
-  }
-
-  PetscFunctionReturn(0);
-}
-
-/* -------------------------------------------------------------------------- */
-PetscErrorCode AllocateMemoryForSubcells(
-  PetscInt    num_cells,
-  PetscInt    num_subcells_per_cell,
-  TDySubcellType subcell_type,
-  TDySubcell    *subcells) {
-
-  PetscFunctionBegin;
-
-  PetscInt num_nu_vectors = GetNumOfNuVectorsForSubcellType(subcell_type);
-  PetscInt num_vertices   = GetNumVerticesForSubcellType(subcell_type);
-  PetscInt num_faces      = GetNumFacesForSubcellType(subcell_type);
-
-  PetscErrorCode ierr;
-  ierr = TDyAllocate_TDyVector_1D(    num_cells*num_subcells_per_cell*num_nu_vectors, &subcells->nu_vector                      ); CHKERRQ(ierr);
-  ierr = TDyAllocate_TDyVector_1D(    num_cells*num_subcells_per_cell*num_nu_vectors, &subcells->nu_star_vector                      ); CHKERRQ(ierr);
-  ierr = TDyAllocate_TDyCoordinate_1D(num_cells*num_subcells_per_cell*num_nu_vectors, &subcells->variable_continuity_coordinates); CHKERRQ(ierr);
-  ierr = TDyAllocate_TDyCoordinate_1D(num_cells*num_subcells_per_cell*num_nu_vectors, &subcells->face_centroid                  ); CHKERRQ(ierr);
-  ierr = TDyAllocate_TDyCoordinate_1D(num_cells*num_subcells_per_cell*num_vertices,   &subcells->vertices_coordinates           ); CHKERRQ(ierr);
-
-  ierr = TDyAllocate_IntegerArray_1D(&subcells->id,num_cells*num_subcells_per_cell          ); CHKERRQ(ierr);
-  ierr = TDyAllocate_IntegerArray_1D(&subcells->num_nu_vectors,num_cells*num_subcells_per_cell          ); CHKERRQ(ierr);
-  ierr = TDyAllocate_IntegerArray_1D(&subcells->num_vertices,num_cells*num_subcells_per_cell          ); CHKERRQ(ierr);
-  ierr = TDyAllocate_IntegerArray_1D(&subcells->num_faces,num_cells*num_subcells_per_cell          ); CHKERRQ(ierr);
-  subcells->type = (TDySubcellType *)malloc(num_cells*num_subcells_per_cell*sizeof(TDySubcellType));
-
-  ierr = TDyAllocate_IntegerArray_1D(&subcells->nu_vector_offset,num_cells*num_subcells_per_cell+1); CHKERRQ(ierr);
-  ierr = TDyAllocate_IntegerArray_1D(&subcells->vertex_offset  ,num_cells*num_subcells_per_cell+1); CHKERRQ(ierr);
-  ierr = TDyAllocate_IntegerArray_1D(&subcells->face_offset     ,num_cells*num_subcells_per_cell+1); CHKERRQ(ierr);
-
-  ierr = TDyAllocate_IntegerArray_1D(&subcells->face_ids        ,num_cells*num_subcells_per_cell*num_faces); CHKERRQ(ierr);
-  ierr = TDyAllocate_IntegerArray_1D(&subcells->is_face_up      ,num_cells*num_subcells_per_cell*num_faces); CHKERRQ(ierr);
-  ierr = TDyAllocate_IntegerArray_1D(&subcells->face_unknown_idx,num_cells*num_subcells_per_cell*num_faces); CHKERRQ(ierr);
-  ierr = TDyAllocate_IntegerArray_1D(&subcells->face_flux_idx   ,num_cells*num_subcells_per_cell*num_faces); CHKERRQ(ierr);
-  ierr = TDyAllocate_RealArray_1D(   &subcells->face_area       ,num_cells*num_subcells_per_cell*num_faces); CHKERRQ(ierr);
-  ierr = TDyAllocate_IntegerArray_1D(&subcells->vertex_ids      ,num_cells*num_subcells_per_cell*num_faces); CHKERRQ(ierr);
-
-  ierr = TDyAllocate_RealArray_1D(   &subcells->T               ,num_cells*num_subcells_per_cell           ); CHKERRQ(ierr);
-
-  for (PetscInt isubcell=0; isubcell<num_cells*num_subcells_per_cell; isubcell++) {
-    subcells->id[isubcell]             = isubcell;
-    subcells->type[isubcell]           = subcell_type;
-
-    subcells->num_nu_vectors[isubcell] = num_nu_vectors;
-    subcells->num_vertices[isubcell]   = num_vertices;
-    subcells->num_faces[isubcell]      = num_faces;
-  }
-
-  for (PetscInt isubcell=0; isubcell <= num_cells*num_subcells_per_cell; isubcell++) {
-    subcells->nu_vector_offset[isubcell] = isubcell*num_nu_vectors;
-    subcells->vertex_offset[isubcell] = isubcell*num_vertices;
-    subcells->face_offset[isubcell] = isubcell*num_faces;
-  }
-
-  PetscFunctionReturn(0);
-}
-
-/* -------------------------------------------------------------------------- */
-PetscErrorCode AllocateMemoryForVertices(
-  PetscInt       num_vertices,
-  PetscInt       ncells_per_vertex,
-  PetscInt       nfaces_per_vertex,
-  PetscInt       nedges_per_vertex,
-  TDyCellType    cell_type,
-  TDyVertex     *vertices) {
-
-  PetscFunctionBegin;
-
-  PetscErrorCode ierr;
-  ierr = TDyAllocate_IntegerArray_1D(&vertices->id                ,num_vertices); CHKERRQ(ierr);
-  ierr = TDyAllocate_IntegerArray_1D(&vertices->global_id         ,num_vertices); CHKERRQ(ierr);
-  ierr = TDyAllocate_IntegerArray_1D(&vertices->num_internal_cells,num_vertices); CHKERRQ(ierr);
-  ierr = TDyAllocate_IntegerArray_1D(&vertices->num_edges         ,num_vertices); CHKERRQ(ierr);
-  ierr = TDyAllocate_IntegerArray_1D(&vertices->num_faces         ,num_vertices); CHKERRQ(ierr);
-  ierr = TDyAllocate_IntegerArray_1D(&vertices->num_boundary_faces,num_vertices); CHKERRQ(ierr);
-
-  vertices->is_local = (PetscBool *)malloc(num_vertices*sizeof(PetscBool));
-
-  ierr = TDyAllocate_TDyCoordinate_1D(num_vertices, &vertices->coordinate); CHKERRQ(ierr);
-
-  ierr = TDyAllocate_IntegerArray_1D(&vertices->edge_offset         ,num_vertices+1); CHKERRQ(ierr);
-  ierr = TDyAllocate_IntegerArray_1D(&vertices->face_offset         ,num_vertices+1); CHKERRQ(ierr);
-  ierr = TDyAllocate_IntegerArray_1D(&vertices->internal_cell_offset,num_vertices+1); CHKERRQ(ierr);
-  ierr = TDyAllocate_IntegerArray_1D(&vertices->subcell_offset      ,num_vertices+1); CHKERRQ(ierr);
-  ierr = TDyAllocate_IntegerArray_1D(&vertices->boundary_face_offset,num_vertices+1); CHKERRQ(ierr);
-
-  ierr = TDyAllocate_IntegerArray_1D(&vertices->edge_ids         ,num_vertices*nedges_per_vertex ); CHKERRQ(ierr);
-  ierr = TDyAllocate_IntegerArray_1D(&vertices->face_ids         ,num_vertices*nfaces_per_vertex ); CHKERRQ(ierr);
-  ierr = TDyAllocate_IntegerArray_1D(&vertices->subface_ids      ,num_vertices*nfaces_per_vertex ); CHKERRQ(ierr);
-  ierr = TDyAllocate_IntegerArray_1D(&vertices->internal_cell_ids,num_vertices*ncells_per_vertex ); CHKERRQ(ierr);
-  ierr = TDyAllocate_IntegerArray_1D(&vertices->subcell_ids      ,num_vertices*nfaces_per_vertex ); CHKERRQ(ierr);
-  ierr = TDyAllocate_IntegerArray_1D(&vertices->boundary_face_ids,num_vertices*nfaces_per_vertex ); CHKERRQ(ierr);
-
-  for (PetscInt ivertex=0; ivertex<num_vertices; ivertex++) {
-    vertices->id[ivertex]                 = ivertex;
-    vertices->is_local[ivertex]           = PETSC_FALSE;
-    vertices->num_internal_cells[ivertex] = 0;
-    vertices->num_edges[ivertex]          = 0;
-    vertices->num_faces[ivertex]          = 0;
-    vertices->num_boundary_faces[ivertex] = 0;
-  }
-
-  for (PetscInt ivertex=0; ivertex<=num_vertices; ivertex++) {
-    vertices->edge_offset[ivertex]          = ivertex*nedges_per_vertex;
-    vertices->face_offset[ivertex]          = ivertex*nfaces_per_vertex;
-    vertices->internal_cell_offset[ivertex] = ivertex*ncells_per_vertex;
-    vertices->subcell_offset[ivertex]       = ivertex*nfaces_per_vertex;
-    vertices->boundary_face_offset[ivertex] = ivertex*nfaces_per_vertex;
-
-  }
-
-  TDyInitialize_IntegerArray_1D(vertices->face_ids, num_vertices*nfaces_per_vertex, 0);
-
-  PetscFunctionReturn(0);
-
-}
-
-/* -------------------------------------------------------------------------- */
-PetscErrorCode AllocateMemoryForEdges(
-  PetscInt num_edges,
-  TDyCellType cell_type,
-  TDyEdge *edges) {
-
-  PetscFunctionBegin;
-
-  PetscInt num_cells = GetNumCellsPerEdgeForCellType(cell_type);
-
-  PetscErrorCode ierr;
-  ierr = TDyAllocate_IntegerArray_1D(&edges->id,num_edges); CHKERRQ(ierr);
-  ierr = TDyAllocate_IntegerArray_1D(&edges->global_id,num_edges); CHKERRQ(ierr);
-  ierr = TDyAllocate_IntegerArray_1D(&edges->num_cells,num_edges); CHKERRQ(ierr);
-  ierr = TDyAllocate_IntegerArray_1D(&edges->vertex_ids,num_edges*2); CHKERRQ(ierr);
-
-  edges->is_local = (PetscBool *)malloc(num_edges*sizeof(PetscBool));
-  edges->is_internal = (PetscBool *)malloc(num_edges*sizeof(PetscBool));
-
-  ierr = TDyAllocate_IntegerArray_1D(&edges->cell_offset,num_edges+1); CHKERRQ(ierr);
-  ierr = TDyAllocate_IntegerArray_1D(&edges->cell_ids,num_edges*num_cells); CHKERRQ(ierr);
-
-  ierr = TDyAllocate_TDyCoordinate_1D(num_edges, &edges->centroid); CHKERRQ(ierr);
-  ierr = TDyAllocate_TDyVector_1D(num_edges, &edges->normal); CHKERRQ(ierr);
-
-  ierr = TDyAllocate_RealArray_1D(&edges->length,num_edges); CHKERRQ(ierr);
-
-  for (PetscInt iedge=0; iedge<num_edges; iedge++) {
-    edges->id[iedge] = iedge;
-    edges->is_local[iedge] = PETSC_FALSE;
-  }
-  for (PetscInt iedge=0; iedge<=num_edges; iedge++) {
-    edges->cell_offset[iedge] = iedge*num_cells;
-  }
-
-  PetscFunctionReturn(0);
-}
-
-/* -------------------------------------------------------------------------- */
-PetscErrorCode AllocateMemoryForFaces(
-  PetscInt num_faces,
-  TDyCellType cell_type,
-  TDyFace *faces) {
-
-  PetscFunctionBegin;
-
-  PetscInt num_cells    = GetNumCellsPerFaceForCellType(cell_type);
-  PetscInt num_edges    = GetNumOfEdgesFormingAFaceForCellType(cell_type);
-  PetscInt num_vertices = GetNumOfVerticesFormingAFaceForCellType(cell_type);
-
-  PetscErrorCode ierr;
-  ierr = TDyAllocate_IntegerArray_1D(&faces->id,num_faces); CHKERRQ(ierr);
-  ierr = TDyAllocate_IntegerArray_1D(&faces->num_vertices,num_faces); CHKERRQ(ierr);
-  ierr = TDyAllocate_IntegerArray_1D(&faces->num_edges,num_faces); CHKERRQ(ierr);
-  ierr = TDyAllocate_IntegerArray_1D(&faces->num_cells,num_faces); CHKERRQ(ierr);
-
-  faces->is_local = (PetscBool *)malloc(num_faces*sizeof(PetscBool));
-  faces->is_internal = (PetscBool *)malloc(num_faces*sizeof(PetscBool));
-
-  ierr = TDyAllocate_IntegerArray_1D(&faces->vertex_offset,num_faces+1); CHKERRQ(ierr);
-  ierr = TDyAllocate_IntegerArray_1D(&faces->cell_offset,num_faces+1); CHKERRQ(ierr);
-  ierr = TDyAllocate_IntegerArray_1D(&faces->edge_offset,num_faces+1); CHKERRQ(ierr);
-
-  ierr = TDyAllocate_IntegerArray_1D(&faces->cell_ids,num_faces*num_cells); CHKERRQ(ierr);
-  ierr = TDyAllocate_IntegerArray_1D(&faces->edge_ids,num_faces*num_edges); CHKERRQ(ierr);
-  ierr = TDyAllocate_IntegerArray_1D(&faces->vertex_ids,num_faces*num_vertices); CHKERRQ(ierr);
-
-  ierr = TDyAllocate_RealArray_1D(&faces->area,num_faces); CHKERRQ(ierr);
-  ierr = TDyAllocate_TDyCoordinate_1D(num_faces, &faces->centroid); CHKERRQ(ierr);
-  ierr = TDyAllocate_TDyVector_1D(num_faces, &faces->normal); CHKERRQ(ierr);
-
-  for (PetscInt iface=0; iface<num_faces; iface++) {
-    faces->id[iface] = iface;
-    faces->is_local[iface]    = PETSC_FALSE;
-    faces->is_internal[iface] = PETSC_FALSE;
-
-    faces->num_edges[iface] = num_edges;
-
-    faces->num_cells[iface] = 0;
-    faces->num_vertices[iface] = 0;
-  }
-
-  for (PetscInt iface=0; iface<=num_faces; iface++) {
-    faces->cell_offset[iface] = iface*num_cells;
-    faces->edge_offset[iface] = iface*num_edges;
-    faces->vertex_offset[iface] = iface*num_vertices;
-  }
-
-  PetscFunctionReturn(0);
-}
-
-
-/* -------------------------------------------------------------------------- */
-PetscErrorCode TDyAllocateMemoryForMesh(TDyMPFAO* mpfao, DM dm) {
-  PetscFunctionBegin;
-
-  TDyMesh *mesh = mpfao->mesh;
-  PetscErrorCode ierr;
-
-  // Determine the number of cells, edges, and vertices of the mesh
-  PetscInt c_start, c_end;
-  ierr = DMPlexGetHeightStratum(dm, 0, &c_start, &c_end); CHKERRQ(ierr);
-  PetscInt num_cells = c_end - c_start;
-
-  PetscInt e_start, e_end;
-  ierr = DMPlexGetDepthStratum( dm, 1, &e_start, &e_end); CHKERRQ(ierr);
-  PetscInt num_edges = e_end - e_start;
-
-  PetscInt v_start, v_end;
-  ierr = DMPlexGetDepthStratum( dm, 0, &v_start, &v_end); CHKERRQ(ierr);
-  PetscInt num_vertices = v_end - v_start;
-
-  PetscInt num_faces;
-  PetscInt f_start, f_end;
-  ierr = DMPlexGetDepthStratum( dm, 2, &f_start, &f_end); CHKERRQ(ierr);
-  num_faces = f_end - f_start;
-
-  mesh->num_cells    = num_cells;
-  mesh->num_faces    = num_faces;
-  mesh->num_edges    = num_edges;
-  mesh->num_vertices = num_vertices;
-
-  mpfao->maxClosureSize = 27*4*4;
-  ierr = TDyAllocate_IntegerArray_1D(&mpfao->closureSize, num_cells+num_faces+num_edges+num_vertices); CHKERRQ(ierr);
-  ierr = TDyAllocate_IntegerArray_2D(&mpfao->closure, num_cells+num_faces+num_edges+num_vertices, 2*mpfao->maxClosureSize); CHKERRQ(ierr);
-  ierr = TDySaveClosures(dm, mpfao->closureSize, mpfao->closure, &mpfao->maxClosureSize); CHKERRQ(ierr);
-
-  // compute number of vertices per grid cell
-  PetscInt nverts_per_cell = TDyGetNumberOfCellVerticesWithClosures(dm, mpfao->closureSize, mpfao->closure);
-  TDyCellType cell_type = GetCellType(nverts_per_cell);
-
-  ierr = AllocateMemoryForCells(num_cells, cell_type, &mesh->cells); CHKERRQ(ierr);
-  ierr = AllocateMemoryForEdges(num_edges, cell_type, &mesh->edges); CHKERRQ(ierr);
-  ierr = AllocateMemoryForFaces(num_faces, cell_type, &mesh->faces); CHKERRQ(ierr);
-
-  PetscInt ncells_per_vertex = TDyMaxNumberOfCellsSharingAVertex(dm, mpfao->closureSize, mpfao->closure);
-  PetscInt nfaces_per_vertex = TDyMaxNumberOfFacesSharingAVertex(dm, mpfao->closureSize, mpfao->closure);
-  PetscInt nedges_per_vertex = TDyMaxNumberOfEdgesSharingAVertex(dm, mpfao->closureSize, mpfao->closure);
-  mpfao->ncv = ncells_per_vertex;
-  mpfao->nfv = nfaces_per_vertex;
-
-  ierr = AllocateMemoryForVertices(num_vertices, ncells_per_vertex, nfaces_per_vertex, nedges_per_vertex, cell_type, &mesh->vertices); CHKERRQ(ierr);
-
-  TDySubcellType subcell_type = GetSubcellTypeForCellType(cell_type);
-  PetscInt num_subcells  = GetNumSubcellsForSubcellType(subcell_type);
-  mesh->num_subcells = num_cells*num_subcells;
-  ierr = AllocateMemoryForSubcells(num_cells, num_subcells, subcell_type, &mesh->subcells); CHKERRQ(ierr);
-
-  ierr = TDyRegionCreate(&mesh->region_connected); CHKERRQ(ierr);
-
-  PetscFunctionReturn(0);
 }
 
 /* -------------------------------------------------------------------------- */
@@ -460,311 +126,6 @@ PetscErrorCode SaveNaturalIDs(TDyMPFAO *mpfao, DM dm){
       ierr = VecDestroy(&region_id_nat_idx); CHKERRQ(ierr);
       ierr = VecDestroy(&global); CHKERRQ(ierr);
       ierr = VecDestroy(&local); CHKERRQ(ierr);
-    }
-  }
-
-  PetscFunctionReturn(0);
-}
-
-/* -------------------------------------------------------------------------- */
-PetscErrorCode SaveMeshGeometricAttributes(TDyMPFAO *mpfao, DM dm) {
-
-  PetscFunctionBegin;
-
-  TDyMesh       *mesh = mpfao->mesh;
-  TDyCell       *cells = &mesh->cells;
-  TDyVertex     *vertices = &mesh->vertices;
-  TDyEdge       *edges = &mesh->edges;
-  TDyFace       *faces = &mesh->faces;
-  PetscErrorCode ierr;
-
-  // Determine the number of cells, edges, and vertices of the mesh
-  PetscInt c_start, c_end;
-  ierr = DMPlexGetHeightStratum(dm, 0, &c_start, &c_end); CHKERRQ(ierr);
-
-  PetscInt e_start, e_end;
-  ierr = DMPlexGetDepthStratum(dm, 1, &e_start, &e_end); CHKERRQ(ierr);
-
-  PetscInt v_start, v_end;
-  ierr = DMPlexGetDepthStratum(dm, 0, &v_start, &v_end); CHKERRQ(ierr);
-
-  PetscInt p_start, pEnd;
-  ierr = DMPlexGetChart(dm, &p_start, &pEnd); CHKERRQ(ierr);
-
-  // Face indexes -- only relevant for 3D calculations.
-  PetscInt dim;
-  ierr = DMGetDimension(dm, &dim); CHKERRQ(ierr);
-
-  PetscInt f_start, f_end;
-  ierr = DMPlexGetDepthStratum( dm, 2, &f_start, &f_end); CHKERRQ(ierr);
-
-  for (PetscInt ielement=p_start; ielement<pEnd; ielement++) {
-
-    if (IsClosureWithinBounds(ielement, v_start, v_end)) { // is the element a vertex?
-      PetscInt ivertex = ielement - v_start;
-      for (PetscInt d=0; d<dim; d++) {
-        vertices->coordinate[ivertex].X[d] = mpfao->X[ielement*dim + d];
-      }
-    } else if (IsClosureWithinBounds(ielement, e_start,
-                                     e_end)) { // is the element an edge?
-      PetscInt iedge = ielement - e_start;
-      for (PetscInt d=0; d<dim; d++) {
-        edges->centroid[iedge].X[d] = mpfao->X[ielement*dim + d];
-        edges->normal[iedge].V[d]   = mpfao->N[ielement*dim + d];
-      }
-    } else if (IsClosureWithinBounds(ielement, c_start,
-                                     c_end)) { // is the element a cell?
-      PetscInt icell = ielement - c_start;
-      cells->volume[icell] = mpfao->V[ielement];
-      for (PetscInt d=0; d<dim; d++) {
-        cells->centroid[icell].X[d] = mpfao->X[ielement*dim + d];
-      }
-    } else if (IsClosureWithinBounds(ielement, f_start,
-                                     f_end)) { // is the elment a face?
-      PetscInt iface = ielement - f_start;
-      for (PetscInt d=0; d<dim; d++) {
-        faces->centroid[iface].X[d] = mpfao->X[ielement*dim + d];
-      }
-      faces->area[iface] = mpfao->V[ielement];
-    }
-  }
-
-  PetscFunctionReturn(0);
-}
-
-/* -------------------------------------------------------------------------- */
-
-PetscErrorCode SaveMeshConnectivityInfo(TDyMPFAO *mpfao, DM dm) {
-
-  PetscFunctionBegin;
-
-  TDyMesh       *mesh = mpfao->mesh;
-  TDyCell       *cells = &mesh->cells;
-  TDyVertex     *vertices = &mesh->vertices;
-  TDyEdge       *edges = &mesh->edges;
-  TDyFace       *faces = &mesh->faces;
-  PetscErrorCode ierr;
-
-  PetscInt nverts_per_cell = mpfao->ncv;
-
-  // Determine the number of cells, edges, and vertices of the mesh
-  PetscInt c_start, c_end;
-  ierr = DMPlexGetHeightStratum(dm, 0, &c_start, &c_end); CHKERRQ(ierr);
-
-  PetscInt e_start, e_end;
-  ierr = DMPlexGetDepthStratum( dm, 1, &e_start, &e_end); CHKERRQ(ierr);
-
-  PetscInt v_start, v_end;
-  ierr = DMPlexGetDepthStratum( dm, 0, &v_start, &v_end); CHKERRQ(ierr);
-
-  PetscInt p_start, pEnd;
-  ierr = DMPlexGetChart(dm, &p_start, &pEnd); CHKERRQ(ierr);
-
-  // Faces -- only relevant in 3D calculations.
-  PetscInt dim, f_start, f_end;
-  ierr = DMGetDimension(dm, &dim); CHKERRQ(ierr);
-  ierr = DMPlexGetDepthStratum( dm, 2, &f_start, &f_end); CHKERRQ(ierr);
-
-  // cell--to--vertex
-  // edge--to--cell
-  // cell--to--edge
-  // edge--to--cell
-  for (PetscInt icell=c_start; icell<c_end; icell++) {
-
-    PetscInt c2v_count, c2e_count, c2f_count;
-    c2v_count = 0;
-    c2e_count = 0;
-    c2f_count = 0;
-
-    for (PetscInt i=0; i<mpfao->closureSize[icell]*2; i+=2)  {
-
-      if (IsClosureWithinBounds(mpfao->closure[icell][i], v_start,
-                                v_end)) { /* Is the closure a vertex? */
-        PetscInt ivertex = mpfao->closure[icell][i] - v_start;
-        PetscInt cOffsetVert = cells->vertex_offset[icell];
-        cells->vertex_ids[cOffsetVert + c2v_count] = ivertex ;
-
-        PetscInt vOffsetCell = vertices->internal_cell_offset[ivertex];
-        PetscInt vOffsetSubcell = vertices->subcell_offset[ivertex];
-
-        PetscInt found = PETSC_FALSE;
-        for (PetscInt j=0; j<nverts_per_cell; j++) {
-          if (vertices->internal_cell_ids[vOffsetCell + j] == -1) {
-            vertices->num_internal_cells[ivertex]++;
-            vertices->internal_cell_ids[vOffsetCell + j] = icell;
-            vertices->subcell_ids[vOffsetSubcell + j]    = c2v_count;
-            found = PETSC_TRUE;
-            break;
-          }
-        }
-        if (!found) {
-          SETERRQ(PETSC_COMM_WORLD,PETSC_ERR_SUP,
-                  "No empty space found in the vertex to save cell");
-        }
-        c2v_count++;
-      } else if (IsClosureWithinBounds(mpfao->closure[icell][i], e_start,
-                                       e_end)) { /* Is the closure an edge? */
-        PetscInt iedge = mpfao->closure[icell][i] - e_start;
-        PetscInt cOffsetEdge = cells->edge_offset[icell];
-        cells->edge_ids[cOffsetEdge + c2e_count] = iedge;
-        PetscInt eOffsetCell = edges->cell_offset[iedge];
-        for (PetscInt j=0; j<2; j++) {
-          if (edges->cell_ids[eOffsetCell + j] == -1) {
-            edges->cell_ids[eOffsetCell + j] = icell;
-            break;
-          }
-        }
-
-        c2e_count++;
-      } else if (IsClosureWithinBounds(mpfao->closure[icell][i], f_start,
-                                       f_end)) { /* Is the closure a face? */
-        PetscInt iface = mpfao->closure[icell][i] - f_start;
-        PetscInt cOffsetFace = cells->face_offset[icell];
-        PetscInt fOffsetCell = faces->cell_offset[iface];
-        cells->face_ids[cOffsetFace + c2f_count] = iface;
-        for (PetscInt j=0; j<2; j++) {
-          if (faces->cell_ids[fOffsetCell + j] < 0) {
-            faces->cell_ids[fOffsetCell + j] = icell;
-            faces->num_cells[iface]++;
-            break;
-          }
-        }
-        c2f_count++;
-      }
-    }
-
-  }
-
-
-  // edge--to--vertex
-  for (PetscInt e=e_start; e<e_end; e++) {
-    const PetscInt* cone;
-    PetscInt cone_size;
-    ierr = DMPlexGetConeSize(dm, e, &cone_size); CHKERRQ(ierr);
-    ierr = DMPlexGetCone(dm, e, &cone); CHKERRQ(ierr);
-
-    PetscInt support_size;
-    ierr = DMPlexGetSupportSize(dm, e, &support_size); CHKERRQ(ierr);
-    PetscInt iedge = e-e_start;
-
-    if (support_size == 1) {
-      edges->is_internal[iedge] = PETSC_FALSE;
-    } else {
-      edges->is_internal[iedge] = PETSC_TRUE;
-    }
-
-    edges->vertex_ids[iedge*2 + 0] = cone[0]-v_start;
-    edges->vertex_ids[iedge*2 + 1] = cone[1]-v_start;
-
-    PetscReal v_1[3], v_2[3];
-    ierr = TDyVertex_GetCoordinate(vertices, edges->vertex_ids[iedge*2 + 0], dim, &v_1[0]); CHKERRQ(ierr);
-    ierr = TDyVertex_GetCoordinate(vertices, edges->vertex_ids[iedge*2 + 1], dim, &v_2[0]); CHKERRQ(ierr);
-
-    for (PetscInt d=0; d<dim; d++) {
-      edges->centroid[iedge].X[d] = (v_1[d] + v_2[d])/2.0;
-    }
-
-    ierr = TDyComputeLength(v_1, v_2, dim, &(edges->length[iedge])); CHKERRQ(ierr);
-  }
-
-  // vertex--to--edge
-  for (PetscInt v=v_start; v<v_end; v++) {
-    const PetscInt *support;
-    PetscInt support_size;
-    ierr = DMPlexGetSupport(dm, v, &support); CHKERRQ(ierr);
-    ierr = DMPlexGetSupportSize(dm, v, &support_size); CHKERRQ(ierr);
-    PetscInt ivertex = v - v_start;
-    vertices->num_edges[ivertex] = support_size;
-    for (PetscInt s=0; s<support_size; s++) {
-      PetscInt iedge = support[s] - e_start;
-      PetscInt vOffsetEdge = vertices->edge_offset[ivertex];
-      vertices->edge_ids[vOffsetEdge + s] = iedge;
-      if (!edges->is_internal[iedge]) vertices->num_boundary_faces[ivertex]++;
-    }
-  }
-
-  for (PetscInt f=f_start; f<f_end; f++){
-    PetscInt iface = f-f_start;
-    PetscInt fOffsetEdge = faces->edge_offset[iface];
-    PetscInt fOffsetVertex = faces->vertex_offset[iface];
-
-    // face--to--edge
-    const PetscInt* cone;
-    PetscInt cone_size;
-    ierr = DMPlexGetConeSize(dm, f, &cone_size); CHKERRQ(ierr);
-    ierr = DMPlexGetCone(dm, f, &cone); CHKERRQ(ierr);
-
-    for (PetscInt c=0;c<cone_size;c++) {
-      faces->edge_ids[fOffsetEdge + c] = cone[c]-e_start;
-    }
-
-    // face--to-vertex
-    for (PetscInt i=0; i<mpfao->closureSize[f]*2; i+=2)  {
-      if (IsClosureWithinBounds(mpfao->closure[f][i],v_start,v_end)) {
-        faces->vertex_ids[fOffsetVertex + faces->num_vertices[iface]] = mpfao->closure[f][i]-v_start;
-        faces->num_vertices[iface]++;
-
-        PetscBool found = PETSC_FALSE;
-        PetscInt ivertex = mpfao->closure[f][i]-v_start;
-        PetscInt vOffsetFace = vertices->face_offset[ivertex];
-        for (PetscInt ii=0; ii<vertices->num_faces[ivertex]; ii++) {
-          if (vertices->face_ids[vOffsetFace+ii] == iface) {
-            found = PETSC_TRUE;
-            break;
-          }
-        }
-        if (!found) {
-          vertices->face_ids[vOffsetFace+vertices->num_faces[ivertex]] = iface;
-          vertices->subface_ids[vOffsetFace+vertices->num_faces[ivertex]] = faces->num_vertices[iface] - 1;
-          vertices->num_faces[ivertex]++;
-        }
-      }
-    }
-
-    // face--to--cell
-    const PetscInt *support;
-    PetscInt support_size;
-    ierr = DMPlexGetSupportSize(dm, f, &support_size); CHKERRQ(ierr);
-    ierr = DMPlexGetSupport(dm, f, &support); CHKERRQ(ierr);
-
-    // TODO: This is where we decide whether a face belongs to the domain
-    // TODO: boundary. It's logically consistent with the way that DMPlex
-    // TODO: decides on the domain boundary, so we can leave it like this
-    // TODO: for now, but we should favor the use of the "boundary" DMLabel
-    // TODO: in future efforts.
-    if (support_size == 2) {
-      faces->is_internal[iface] = PETSC_TRUE;
-    } else {
-      faces->is_internal[iface] = PETSC_FALSE;
-    }
-
-    for (PetscInt s=0; s<support_size; s++) {
-      PetscInt icell = support[s] - c_start;
-      PetscBool found = PETSC_FALSE;
-      PetscInt cOffsetFace = cells->face_offset[icell];
-      for (PetscInt ii=0; ii<cells->num_faces[icell]; ii++) {
-        if (cells->face_ids[cOffsetFace+ii] == f-f_start) {
-          found = PETSC_TRUE;
-          break;
-        }
-      }
-      if (!found) {
-        cells->face_ids[cOffsetFace + cells->num_faces[icell]] = f-f_start;
-        cells->num_faces[icell]++;
-        found = PETSC_TRUE;
-      }
-    }
-
-    // If it is a boundary face, increment the number of boundary
-    // cells by 1 for all vertices that form the face
-    if (!faces->is_internal[iface]) {
-      for (PetscInt v=0; v<faces->num_vertices[iface]; v++) {
-        PetscInt vertex_id = faces->vertex_ids[fOffsetVertex + v];
-        PetscInt vOffsetBoundaryFace = vertices->boundary_face_offset[vertex_id];
-
-        vertices->boundary_face_ids[vOffsetBoundaryFace + vertices->num_boundary_faces[vertex_id] ] = iface;
-        vertices->num_boundary_faces[vertex_id]++;
-      }
     }
   }
 
@@ -3242,378 +2603,3 @@ PetscErrorCode UpdateCellOrientationAroundAFace(TDy tdy) {
   PetscFunctionReturn(0);
 }
 
-/* -------------------------------------------------------------------------- */
-/// Reads a geometric attribue file. Currently only PETSc binary file format
-/// is supported
-///
-/// @param [inout] tdy A TDy struct
-/// @returns 0  on success or a non-zero error code on failure
-PetscErrorCode ReadMeshGeometricAttributes(TDy tdy) {
-  PetscFunctionBegin;
-  PetscErrorCode ierr;
-
-  TDyMesh *mesh = tdy->mesh;
-  TDyCell *cells = &mesh->cells;
-  TDyFace *faces = &mesh->faces;
-
-  PetscInt rank;
-  MPI_Comm_rank(PETSC_COMM_WORLD,&rank);
-
-  // 1. Read the binary geometric attribute file
-  //
-  // The i-th entry of the output strided Vec is
-  //
-  // volume_i, centroid_<x|y|z>_i, numNeighbor, nat_id_ij, face_area_of_ij, face_centroid_<x|y|z>_ij
-  //
-  
-  PetscInt numLocalCells = TDyMeshGetNumberOfLocalCells(mesh);
-  PetscInt numNeighbor = cells->num_neighbors[0]; // It is assumed that all cells have same number of neighbors
-  PetscInt numCellAttr = 5; // volume_i, centroid_<x|y|z>_i, numNeighbor
-  PetscInt numNeigbhorAttr = 5; // nat_id_ij, face_centroid_<x|y|z>_ij face_area_of_ij
-  PetscInt stride = numCellAttr + numNeighbor * numNeigbhorAttr;
-
-  Vec natural_vec;
-  ierr = VecCreate(PETSC_COMM_WORLD,&natural_vec); CHKERRQ(ierr);
-  ierr = VecSetSizes(natural_vec,numLocalCells*stride,PETSC_DECIDE); CHKERRQ(ierr);
-  ierr = VecSetBlockSize(natural_vec,stride); CHKERRQ(ierr);
-  ierr = VecSetFromOptions(natural_vec); CHKERRQ(ierr);
-
-  Vec local_vec;
-  ierr = VecCreate(PETSC_COMM_WORLD,&local_vec); CHKERRQ(ierr);
-  ierr = VecSetSizes(local_vec,mesh->num_cells*stride,PETSC_DECIDE); CHKERRQ(ierr);
-  ierr = VecSetBlockSize(local_vec,stride); CHKERRQ(ierr);
-  ierr = VecSetFromOptions(local_vec); CHKERRQ(ierr);
-
-  // Read the data
-  ierr = TDyReadBinaryPetscVec(natural_vec,tdy->options.geom_attributes_file); CHKERRQ(ierr);
-
-  //
-  // 2. Perform natural-to-local scatter of the geometric attributes
-  //
-
-  // Create index for natural vector
-  PetscInt local_offset = 0;
-  ierr = MPI_Exscan(&mesh->num_cells,&local_offset,1,MPI_INTEGER,MPI_SUM,PETSC_COMM_WORLD);
-
-  PetscInt int_array_from[mesh->num_cells];
-  PetscInt int_array_to[mesh->num_cells];
-  for (PetscInt i=0; i<mesh->num_cells; i++){
-    int_array_to[i] = i + local_offset;
-    int_array_from[i] = cells->natural_id[i];
-  }
-
-  IS is_from;
-  ierr = ISCreateBlock(PETSC_COMM_WORLD,stride,mesh->num_cells,int_array_from,PETSC_COPY_VALUES,&is_from); CHKERRQ(ierr);
-
-  // Create index for local vector
-  IS is_to;
-  ierr = ISCreateBlock(PETSC_COMM_WORLD,stride,mesh->num_cells,int_array_to,PETSC_COPY_VALUES,&is_to); CHKERRQ(ierr);
-
-  // Create natural-to-local scatter
-  VecScatter n2l_scatter;
-  VecScatterCreate(natural_vec,is_from,local_vec,is_to,&n2l_scatter); CHKERRQ(ierr);
-
-  ierr = VecScatterBegin(n2l_scatter,natural_vec,local_vec,INSERT_VALUES,SCATTER_FORWARD); CHKERRQ(ierr);
-  ierr = VecScatterEnd(n2l_scatter,natural_vec,local_vec,INSERT_VALUES,SCATTER_FORWARD); CHKERRQ(ierr);
-  TDySavePetscVecAsBinary(local_vec,"cells_local.bin");
-
-  //
-  // 3. Save geometric attributes locally
-  //
-  PetscScalar *values;
-  ierr = VecGetArray(local_vec,&values); CHKERRQ(ierr);
-
-  for (PetscInt icell=0; icell<mesh->num_cells; icell++){
-
-    // The i-th entry of the output strided Vec is
-    //
-    //  col 1        col 2-4            col-5       col 6         col 7            col 8-10
-    // volume_i, centroid_<x|y|z>_i, numNeighbor, nat_id_ij, face_area_of_ij, face_centroid_<x|y|z>_ij, ...
-
-    PetscInt col = 0;
-    PetscInt offset = icell*stride;
-
-    // col 1: cell volume
-    cells->volume[icell] = values[ offset + col];      
-    col++;
-
-    // col 2-4: cell centroid
-    for (PetscInt d=0; d<3; d++) {
-      cells->centroid[icell].X[d] = values[offset + col];
-      col++;
-    }
-
-    PetscInt *face_ids, num_faces;
-    TDyMeshGetCellFaces(mesh, icell, &face_ids, &num_faces);
-
-    // col 5: number of neighbors
-    if ( num_faces != values[offset + col]) {
-      printf("ReadMeshGeometricAttributes: Number of faces do not match = num_faces%d values[%d] = %f; rank = %d\n",num_faces, offset + col, values[offset + col],rank);
-      SETERRQ(PETSC_COMM_SELF, PETSC_ERR_LIB, "ReadMeshGeometricAttributes: Number of faces do not match");
-    }
-    col++;
-
-    // Initialize if information about face/neigbhor is set
-    PetscInt bnd_face_set[num_faces];
-    for (PetscInt iface=0; iface<num_faces; iface++) {
-      bnd_face_set[iface] = 0;
-    }
-
-    // col 6-10; 11-15; ...
-    for (PetscInt iface=0; iface<num_faces; iface++) {
-
-      PetscBool face_found = PETSC_FALSE;
-      PetscInt face_id;
-
-      // col 6
-      PetscInt neighbor_id_in_data = values[offset + col];
-      col++;
-
-      if (neighbor_id_in_data < 0) {
-
-        // This face/neighbor is on boundary.
-        // Find the first boundary face for which geometric attributes
-        // hasn't already been set
-
-        for (PetscInt i=0; i<num_faces; i++) {
-          face_id = face_ids[i];
-
-          PetscInt *cell_ids, tmp;
-          ierr = TDyMeshGetFaceCells(mesh, face_id, &cell_ids, &tmp); CHKERRQ(ierr);
-          if (bnd_face_set[i] == 0 && (cell_ids[0]<0 || cell_ids[1]<0)) {
-            bnd_face_set[i] = 1;
-            face_found = PETSC_TRUE;
-            break;
-          }
-        }
-      } else {
-
-        // This face/neigbhor is internal. 
-        // Loop over all the faces of the cell to find the face index of
-        // the current cell whose neighbor's natural id is neighbor_id_in_data
-
-        for (PetscInt i=0; i<num_faces; i++) {
-          face_id = face_ids[i];
-          PetscInt *cell_ids, tmp;
-          ierr = TDyMeshGetFaceCells(mesh, face_id, &cell_ids, &tmp); CHKERRQ(ierr);
-
-          PetscInt nat_id_up = cells->natural_id[cell_ids[0]];
-          PetscInt nat_id_dn = cells->natural_id[cell_ids[1]];
-
-          if ( nat_id_up == neighbor_id_in_data || nat_id_dn == neighbor_id_in_data){
-            face_found = PETSC_TRUE;
-            break;
-          }
-        }
-      }
-
-      if (!face_found && cells->is_local[icell]) {
-        // Only stop if the cell is internal and a face is not found
-        printf("Not found icell == %d; iface == %d; neighbor_id_in_data = %d; rank = %d\n",icell,iface,neighbor_id_in_data,rank);
-        SETERRQ(PETSC_COMM_SELF, PETSC_ERR_LIB, "ReadMeshGeometricAttributes: Face not found");
-      } else {
-
-        // col 7: Save face area
-        faces->area[face_id] = values[offset + col];
-        col++;
-
-        // col 8-10: Save face centroid
-        for (PetscInt d=0; d<3; d++) {
-          faces->centroid[face_id].X[d] = values[offset + col];
-          col++;
-        }
-      }
-    }    
-  }
-  ierr = VecRestoreArray(local_vec,&values); CHKERRQ(ierr);
-
-  PetscFunctionReturn(0);
-}
-
-/* -------------------------------------------------------------------------- */
-/// Outputs geometric attribues of mesh to a file. Currently only PETSc binary
-/// file format is supported.
-///
-/// @param [inout] tdy A TDy struct
-/// @returns 0  on success or a non-zero error code on failure
-PetscErrorCode OutputMeshGeometricAttributes(TDy tdy) {
-  PetscFunctionBegin;
-  PetscErrorCode ierr;
-
-  TDyMesh *mesh = tdy->mesh;
-  TDyCell *cells = &mesh->cells;
-  TDyFace *faces = &mesh->faces;
-
-  // The i-th entry of the output strided Vec is
-  //
-    //  col 1        col 2-4            col-5       col 6         col 7            col 8-10
-  // volume_i, centroid_<x|y|z>_i, numNeighbor, nat_id_ij, face_area_of_ij, face_centroid_<x|y|z>_ij
-  //
-
-  PetscInt numLocalCells = TDyMeshGetNumberOfLocalCells(mesh);
-
-  PetscInt numNeighbor = cells->num_neighbors[0]; // It is assumed that all cells have same number of neighbors
-  PetscInt numCellAttr = 5; // volume_i, centroid_<x|y|z>_i, numNeighbor
-  PetscInt numNeigbhorAttr = 5; // nat_id_ij, face_centroid_<x|y|z>_ij face_area_of_ij
-  PetscInt stride = numCellAttr + numNeighbor * numNeigbhorAttr;
-
-  // Creates vectors for storing geometric attributes in global and natural
-  // index
-  Vec global_vec, natural_vec;
-  ierr = VecCreate(PETSC_COMM_WORLD,&global_vec); CHKERRQ(ierr);
-  ierr = VecSetSizes(global_vec,numLocalCells*stride,PETSC_DECIDE); CHKERRQ(ierr);
-  ierr = VecSetBlockSize(global_vec,stride); CHKERRQ(ierr);
-  ierr = VecSetFromOptions(global_vec); CHKERRQ(ierr);
-
-  ierr = VecDuplicate(global_vec, &natural_vec); CHKERRQ(ierr);
-
-  PetscScalar *values;
-  ierr = VecGetArray(global_vec,&values); CHKERRQ(ierr);
-
-  // Fill in the geometric attributes for only local cells
-  PetscInt count=0;
-  for (PetscInt icell=0; icell<mesh->num_cells; icell++){
-    if (mesh->cells.is_local[icell]) {
-      PetscInt offset = 0;
-
-      // cell volume
-      values[count*stride + offset] = cells->volume[icell];      
-      offset++;
-
-      // cell centroid
-      for (PetscInt d=0; d<3; d++) {
-        values[count*stride + offset] = cells->centroid[icell].X[d];
-        offset++;
-      }
-
-      PetscInt *face_ids, num_faces;
-      TDyMeshGetCellFaces(mesh, icell, &face_ids, &num_faces);
-
-      // number of neighbors
-      values[count*stride + offset] = num_faces;
-      offset++;
-
-      for (PetscInt n=0; n<num_faces; n++){
-        PetscInt face_id = face_ids[n];
-
-        PetscInt *face_cell_ids, tmp;
-        ierr = TDyMeshGetFaceCells(mesh, face_id, &face_cell_ids, &tmp); CHKERRQ(ierr);
-
-        PetscInt neighbor_id;
-        if (face_cell_ids[0] == icell) {
-          neighbor_id = face_cell_ids[1];
-        } else {
-          neighbor_id = face_cell_ids[0];
-        }
-
-        PetscInt neighbor_nat_id = neighbor_id;
-        
-        if (neighbor_id >= 0) {
-          neighbor_nat_id = cells->natural_id[neighbor_id];
-        }
-
-        // natural ID of neigbhor
-        values[count*stride + offset] = neighbor_nat_id;
-        offset++;
-
-        // face area
-        values[count*stride + offset] = faces->area[face_id];
-        offset++;
-
-        // face centroid
-        for (PetscInt d=0; d<3; d++) {
-          values[count*stride + offset] = faces->centroid[face_id].X[d];
-          offset++;
-        }
-
-      } // loop over neighbors
-
-    count++;
-    } // loop over local cells
-  }
-  ierr = VecRestoreArray(global_vec,&values); CHKERRQ(ierr);
-
-  // Create (from/to) Index Sets for vector scatter
-  PetscInt global_offset = 0;
-  ierr = MPI_Exscan(&numLocalCells,&global_offset,1,MPI_INTEGER,MPI_SUM,PETSC_COMM_WORLD);
-
-  PetscInt int_array[numLocalCells];
-  for (PetscInt i=0; i<numLocalCells; i++){
-    int_array[i] = i + global_offset;
-  }
-
-  IS is_from;
-  ierr = ISCreateBlock(PETSC_COMM_WORLD,stride,numLocalCells,int_array,PETSC_COPY_VALUES,&is_from); CHKERRQ(ierr);
-
-  IS is_to;
-  count = 0;
-  for (PetscInt icell=0; icell<mesh->num_cells; icell++){
-    if (mesh->cells.is_local[icell]) {
-      int_array[count++] = cells->natural_id[icell];
-    }
-  }
-  ierr = ISCreateBlock(PETSC_COMM_WORLD,stride,numLocalCells,int_array,PETSC_COPY_VALUES,&is_to); CHKERRQ(ierr);
-
-  // Create vector scatter
-  VecScatter g2n_scatter;
-  VecScatterCreate(global_vec,is_from,natural_vec,is_to,&g2n_scatter); CHKERRQ(ierr);
-
-  // Scatter data
-  ierr = VecScatterBegin(g2n_scatter,global_vec,natural_vec,INSERT_VALUES,SCATTER_FORWARD); CHKERRQ(ierr);
-  ierr = VecScatterEnd(g2n_scatter,global_vec,natural_vec,INSERT_VALUES,SCATTER_FORWARD); CHKERRQ(ierr);
-
-  // Output data
-  ierr = TDySavePetscVecAsBinary(natural_vec,tdy->options.geom_attributes_file); CHKERRQ(ierr);
-
-  // Cleanup
-  ierr = ISDestroy(&is_from); CHKERRQ(ierr);
-  ierr = ISDestroy(&is_to); CHKERRQ(ierr);
-  ierr = VecScatterDestroy(&g2n_scatter); CHKERRQ(ierr);
-
-  PetscFunctionReturn(0);
-}
-
-/* -------------------------------------------------------------------------- */
-PetscErrorCode TDyBuildMesh(TDyMPFAO *mpfao, DM dm) {
-  PetscFunctionBegin;
-  TDY_START_FUNCTION_TIMER()
-  PetscErrorCode ierr;
-
-  PetscInt dim;
-  ierr = DMGetDimension(dm, &dim); CHKERRQ(ierr);
-
-  if (dim != 3) {
-    SETERRQ(PETSC_COMM_WORLD,PETSC_ERR_USER,"MPFA-O only supports 3D meshes");
-  }
-
-  ierr = SaveMeshConnectivityInfo(mpfao, dm); CHKERRQ(ierr);
-
-  ierr = IdentifyLocalCells(mpfao); CHKERRQ(ierr);
-  ierr = IdentifyLocalVertices(mpfao); CHKERRQ(ierr);
-  ierr = IdentifyLocalEdges(mpfao); CHKERRQ(ierr);
-  ierr = IdentifyLocalFaces(mpfao); CHKERRQ(ierr);
-
-  ierr = SaveNaturalIDs(mpfao, dm); CHKERRQ(ierr);
-  if (mpfao->read_geom_attributes) {
-    ierr = ReadMeshGeometricAttributes(mpfao); CHKERRQ(ierr);
-  } {
-    ierr = SaveMeshGeometricAttributes(mpfao); CHKERRQ(ierr);
-  }
-  mpfao->read_geom_attributes = 0;
-
-  if (mpfao->output_geom_attributes) {
-    ierr = OutputMeshGeometricAttributes(mpfao); CHKERRQ(ierr);
-  }
-  mpfao->output_geom_attributes = 0;
-
-  ierr = ConvertCellsToCompressedFormat(mpfao); CHKERRQ(ierr);
-  ierr = ConvertVerticesToCompressedFormat(mpfao); CHKERRQ(ierr);
-  ierr = ConvertSubcellsToCompressedFormat(mpfao); CHKERRQ(ierr);
-  ierr = ConvertFacesToCompressedFormat(mpfao); CHKERRQ(ierr);
-  ierr = UpdateFaceOrderAroundAVertex(mpfao); CHKERRQ(ierr);
-  ierr = UpdateCellOrientationAroundAFace(mpfao); CHKERRQ(ierr);
-  ierr = SetupSubcells(mpfao); CHKERRQ(ierr);
-
-  TDY_STOP_FUNCTION_TIMER()
-  PetscFunctionReturn(0);
-}
