+++ src/mpfao/tdympfao.c
@@ -348,21 +348,326 @@ static PetscErrorCode TDyMPFAO_AllocateMemoryForEnergySourceSinkValues(TDyMPFAO 
   PetscFunctionReturn(0);
 }
 
-PetscErrorCode TDyMPFAOSetGmatrixMethod(TDyMPFAO* mpfao,
+static PetscErrorCode SaveMeshConnectivityInfo(TDyMPFAO *mpfao, DM dm) {
+
+  PetscFunctionBegin;
+
+  TDyMesh       *mesh = mpfao->mesh;
+  TDyCell       *cells = &mesh->cells;
+  TDyVertex     *vertices = &mesh->vertices;
+  TDyEdge       *edges = &mesh->edges;
+  TDyFace       *faces = &mesh->faces;
+  PetscErrorCode ierr;
+
+  PetscInt nverts_per_cell = mpfao->ncv;
+
+  // Determine the number of cells, edges, and vertices of the mesh
+  PetscInt c_start, c_end;
+  ierr = DMPlexGetHeightStratum(dm, 0, &c_start, &c_end); CHKERRQ(ierr);
+
+  PetscInt e_start, e_end;
+  ierr = DMPlexGetDepthStratum( dm, 1, &e_start, &e_end); CHKERRQ(ierr);
+
+  PetscInt v_start, v_end;
+  ierr = DMPlexGetDepthStratum( dm, 0, &v_start, &v_end); CHKERRQ(ierr);
+
+  PetscInt p_start, pEnd;
+  ierr = DMPlexGetChart(dm, &p_start, &pEnd); CHKERRQ(ierr);
+
+  // Faces -- only relevant in 3D calculations.
+  PetscInt dim, f_start, f_end;
+  ierr = DMGetDimension(dm, &dim); CHKERRQ(ierr);
+  ierr = DMPlexGetDepthStratum( dm, 2, &f_start, &f_end); CHKERRQ(ierr);
+
+  // cell--to--vertex
+  // edge--to--cell
+  // cell--to--edge
+  // edge--to--cell
+  for (PetscInt icell=c_start; icell<c_end; icell++) {
+
+    PetscInt c2v_count, c2e_count, c2f_count;
+    c2v_count = 0;
+    c2e_count = 0;
+    c2f_count = 0;
+
+    for (PetscInt i=0; i<mpfao->closureSize[icell]*2; i+=2)  {
+
+      if (IsClosureWithinBounds(mpfao->closure[icell][i], v_start,
+                                v_end)) { /* Is the closure a vertex? */
+        PetscInt ivertex = mpfao->closure[icell][i] - v_start;
+        PetscInt cOffsetVert = cells->vertex_offset[icell];
+        cells->vertex_ids[cOffsetVert + c2v_count] = ivertex ;
+
+        PetscInt vOffsetCell = vertices->internal_cell_offset[ivertex];
+        PetscInt vOffsetSubcell = vertices->subcell_offset[ivertex];
+
+        PetscInt found = PETSC_FALSE;
+        for (PetscInt j=0; j<nverts_per_cell; j++) {
+          if (vertices->internal_cell_ids[vOffsetCell + j] == -1) {
+            vertices->num_internal_cells[ivertex]++;
+            vertices->internal_cell_ids[vOffsetCell + j] = icell;
+            vertices->subcell_ids[vOffsetSubcell + j]    = c2v_count;
+            found = PETSC_TRUE;
+            break;
+          }
+        }
+        if (!found) {
+          SETERRQ(PETSC_COMM_WORLD,PETSC_ERR_SUP,
+                  "No empty space found in the vertex to save cell");
+        }
+        c2v_count++;
+      } else if (IsClosureWithinBounds(mpfao->closure[icell][i], e_start,
+                                       e_end)) { /* Is the closure an edge? */
+        PetscInt iedge = mpfao->closure[icell][i] - e_start;
+        PetscInt cOffsetEdge = cells->edge_offset[icell];
+        cells->edge_ids[cOffsetEdge + c2e_count] = iedge;
+        PetscInt eOffsetCell = edges->cell_offset[iedge];
+        for (PetscInt j=0; j<2; j++) {
+          if (edges->cell_ids[eOffsetCell + j] == -1) {
+            edges->cell_ids[eOffsetCell + j] = icell;
+            break;
+          }
+        }
+
+        c2e_count++;
+      } else if (IsClosureWithinBounds(mpfao->closure[icell][i], f_start,
+                                       f_end)) { /* Is the closure a face? */
+        PetscInt iface = mpfao->closure[icell][i] - f_start;
+        PetscInt cOffsetFace = cells->face_offset[icell];
+        PetscInt fOffsetCell = faces->cell_offset[iface];
+        cells->face_ids[cOffsetFace + c2f_count] = iface;
+        for (PetscInt j=0; j<2; j++) {
+          if (faces->cell_ids[fOffsetCell + j] < 0) {
+            faces->cell_ids[fOffsetCell + j] = icell;
+            faces->num_cells[iface]++;
+            break;
+          }
+        }
+        c2f_count++;
+      }
+    }
+
+  }
+
+
+  // edge--to--vertex
+  for (PetscInt e=e_start; e<e_end; e++) {
+    const PetscInt* cone;
+    PetscInt cone_size;
+    ierr = DMPlexGetConeSize(dm, e, &cone_size); CHKERRQ(ierr);
+    ierr = DMPlexGetCone(dm, e, &cone); CHKERRQ(ierr);
+
+    PetscInt support_size;
+    ierr = DMPlexGetSupportSize(dm, e, &support_size); CHKERRQ(ierr);
+    PetscInt iedge = e-e_start;
+
+    if (support_size == 1) {
+      edges->is_internal[iedge] = PETSC_FALSE;
+    } else {
+      edges->is_internal[iedge] = PETSC_TRUE;
+    }
+
+    edges->vertex_ids[iedge*2 + 0] = cone[0]-v_start;
+    edges->vertex_ids[iedge*2 + 1] = cone[1]-v_start;
+
+    PetscReal v_1[3], v_2[3];
+    ierr = TDyVertex_GetCoordinate(vertices, edges->vertex_ids[iedge*2 + 0], dim, &v_1[0]); CHKERRQ(ierr);
+    ierr = TDyVertex_GetCoordinate(vertices, edges->vertex_ids[iedge*2 + 1], dim, &v_2[0]); CHKERRQ(ierr);
+
+    for (PetscInt d=0; d<dim; d++) {
+      edges->centroid[iedge].X[d] = (v_1[d] + v_2[d])/2.0;
+    }
+
+    ierr = TDyComputeLength(v_1, v_2, dim, &(edges->length[iedge])); CHKERRQ(ierr);
+  }
+
+  // vertex--to--edge
+  for (PetscInt v=v_start; v<v_end; v++) {
+    const PetscInt *support;
+    PetscInt support_size;
+    ierr = DMPlexGetSupport(dm, v, &support); CHKERRQ(ierr);
+    ierr = DMPlexGetSupportSize(dm, v, &support_size); CHKERRQ(ierr);
+    PetscInt ivertex = v - v_start;
+    vertices->num_edges[ivertex] = support_size;
+    for (PetscInt s=0; s<support_size; s++) {
+      PetscInt iedge = support[s] - e_start;
+      PetscInt vOffsetEdge = vertices->edge_offset[ivertex];
+      vertices->edge_ids[vOffsetEdge + s] = iedge;
+      if (!edges->is_internal[iedge]) vertices->num_boundary_faces[ivertex]++;
+    }
+  }
+
+  for (PetscInt f=f_start; f<f_end; f++){
+    PetscInt iface = f-f_start;
+    PetscInt fOffsetEdge = faces->edge_offset[iface];
+    PetscInt fOffsetVertex = faces->vertex_offset[iface];
+
+    // face--to--edge
+    const PetscInt* cone;
+    PetscInt cone_size;
+    ierr = DMPlexGetConeSize(dm, f, &cone_size); CHKERRQ(ierr);
+    ierr = DMPlexGetCone(dm, f, &cone); CHKERRQ(ierr);
+
+    for (PetscInt c=0;c<cone_size;c++) {
+      faces->edge_ids[fOffsetEdge + c] = cone[c]-e_start;
+    }
+
+    // face--to-vertex
+    for (PetscInt i=0; i<mpfao->closureSize[f]*2; i+=2)  {
+      if (IsClosureWithinBounds(mpfao->closure[f][i],v_start,v_end)) {
+        faces->vertex_ids[fOffsetVertex + faces->num_vertices[iface]] = mpfao->closure[f][i]-v_start;
+        faces->num_vertices[iface]++;
+
+        PetscBool found = PETSC_FALSE;
+        PetscInt ivertex = mpfao->closure[f][i]-v_start;
+        PetscInt vOffsetFace = vertices->face_offset[ivertex];
+        for (PetscInt ii=0; ii<vertices->num_faces[ivertex]; ii++) {
+          if (vertices->face_ids[vOffsetFace+ii] == iface) {
+            found = PETSC_TRUE;
+            break;
+          }
+        }
+        if (!found) {
+          vertices->face_ids[vOffsetFace+vertices->num_faces[ivertex]] = iface;
+          vertices->subface_ids[vOffsetFace+vertices->num_faces[ivertex]] = faces->num_vertices[iface] - 1;
+          vertices->num_faces[ivertex]++;
+        }
+      }
+    }
+
+    // face--to--cell
+    const PetscInt *support;
+    PetscInt support_size;
+    ierr = DMPlexGetSupportSize(dm, f, &support_size); CHKERRQ(ierr);
+    ierr = DMPlexGetSupport(dm, f, &support); CHKERRQ(ierr);
+
+    // TODO: This is where we decide whether a face belongs to the domain
+    // TODO: boundary. It's logically consistent with the way that DMPlex
+    // TODO: decides on the domain boundary, so we can leave it like this
+    // TODO: for now, but we should favor the use of the "boundary" DMLabel
+    // TODO: in future efforts.
+    if (support_size == 2) {
+      faces->is_internal[iface] = PETSC_TRUE;
+    } else {
+      faces->is_internal[iface] = PETSC_FALSE;
+    }
+
+    for (PetscInt s=0; s<support_size; s++) {
+      PetscInt icell = support[s] - c_start;
+      PetscBool found = PETSC_FALSE;
+      PetscInt cOffsetFace = cells->face_offset[icell];
+      for (PetscInt ii=0; ii<cells->num_faces[icell]; ii++) {
+        if (cells->face_ids[cOffsetFace+ii] == f-f_start) {
+          found = PETSC_TRUE;
+          break;
+        }
+      }
+      if (!found) {
+        cells->face_ids[cOffsetFace + cells->num_faces[icell]] = f-f_start;
+        cells->num_faces[icell]++;
+        found = PETSC_TRUE;
+      }
+    }
+
+    // If it is a boundary face, increment the number of boundary
+    // cells by 1 for all vertices that form the face
+    if (!faces->is_internal[iface]) {
+      for (PetscInt v=0; v<faces->num_vertices[iface]; v++) {
+        PetscInt vertex_id = faces->vertex_ids[fOffsetVertex + v];
+        PetscInt vOffsetBoundaryFace = vertices->boundary_face_offset[vertex_id];
+
+        vertices->boundary_face_ids[vOffsetBoundaryFace + vertices->num_boundary_faces[vertex_id] ] = iface;
+        vertices->num_boundary_faces[vertex_id]++;
+      }
+    }
+  }
+
+  PetscFunctionReturn(0);
+}
+
+PetscErrorCode SaveMeshGeometry(TDyMPFAO *mpfao, DM dm) {
+
+  PetscFunctionBegin;
+
+  TDyMesh       *mesh = mpfao->mesh;
+  TDyCell       *cells = &mesh->cells;
+  TDyVertex     *vertices = &mesh->vertices;
+  TDyEdge       *edges = &mesh->edges;
+  TDyFace       *faces = &mesh->faces;
+  PetscErrorCode ierr;
+
+  // Determine the number of cells, edges, and vertices of the mesh
+  PetscInt c_start, c_end;
+  ierr = DMPlexGetHeightStratum(dm, 0, &c_start, &c_end); CHKERRQ(ierr);
+
+  PetscInt e_start, e_end;
+  ierr = DMPlexGetDepthStratum(dm, 1, &e_start, &e_end); CHKERRQ(ierr);
+
+  PetscInt v_start, v_end;
+  ierr = DMPlexGetDepthStratum(dm, 0, &v_start, &v_end); CHKERRQ(ierr);
+
+  PetscInt p_start, pEnd;
+  ierr = DMPlexGetChart(dm, &p_start, &pEnd); CHKERRQ(ierr);
+
+  // Face indexes -- only relevant for 3D calculations.
+  PetscInt dim;
+  ierr = DMGetDimension(dm, &dim); CHKERRQ(ierr);
+
+  PetscInt f_start, f_end;
+  ierr = DMPlexGetDepthStratum( dm, 2, &f_start, &f_end); CHKERRQ(ierr);
+
+  for (PetscInt ielement=p_start; ielement<pEnd; ielement++) {
+
+    if (IsClosureWithinBounds(ielement, v_start, v_end)) { // is the element a vertex?
+      PetscInt ivertex = ielement - v_start;
+      for (PetscInt d=0; d<dim; d++) {
+        vertices->coordinate[ivertex].X[d] = mpfao->X[ielement*dim + d];
+      }
+    } else if (IsClosureWithinBounds(ielement, e_start,
+                                     e_end)) { // is the element an edge?
+      PetscInt iedge = ielement - e_start;
+      for (PetscInt d=0; d<dim; d++) {
+        edges->centroid[iedge].X[d] = mpfao->X[ielement*dim + d];
+        edges->normal[iedge].V[d]   = mpfao->N[ielement*dim + d];
+      }
+    } else if (IsClosureWithinBounds(ielement, c_start,
+                                     c_end)) { // is the element a cell?
+      PetscInt icell = ielement - c_start;
+      cells->volume[icell] = mpfao->V[ielement];
+      for (PetscInt d=0; d<dim; d++) {
+        cells->centroid[icell].X[d] = mpfao->X[ielement*dim + d];
+      }
+    } else if (IsClosureWithinBounds(ielement, f_start,
+                                     f_end)) { // is the elment a face?
+      PetscInt iface = ielement - f_start;
+      for (PetscInt d=0; d<dim; d++) {
+        faces->centroid[iface].X[d] = mpfao->X[ielement*dim + d];
+      }
+      faces->area[iface] = mpfao->V[ielement];
+    }
+  }
+
+  PetscFunctionReturn(0);
+}
+
+
+PetscErrorCode TDyMPFAOSetGmatrixMethod(TDy tdy,
                                         TDyMPFAOGmatrixMethod method) {
   PetscFunctionBegin;
 
   PetscValidPointer(tdy,1);
+  TDyMPFAO *mpfao = tdy->context;
   mpfao->gmatrix_method = method;
 
   PetscFunctionReturn(0);
 }
 
-PetscErrorCode TDyMPFAOSetBoundaryConditionType(TDyMPFAO* mpfao,
+PetscErrorCode TDyMPFAOSetBoundaryConditionType(TDy tdy,
                                                 TDyMPFAOBoundaryConditionType bctype) {
   PetscFunctionBegin;
 
   PetscValidPointer(tdy,1);
+  TDyMPFAO *mpfao = tdy->context;
   mpfao->bc_type = bctype;
 
   PetscFunctionReturn(0);
@@ -611,9 +916,23 @@ static PetscErrorCode TDySetup_Common(TDyMPFAO *mpfao,
     MaterialPropComputeSoilSpecificHeat(matprop, nc, mpfao->X, mpfao->c_soil);
   }
 
-  // Allocate mesh data.
-  mpfao->mesh = (TDyMesh *) malloc(sizeof(TDyMesh));
-  ierr = TDyAllocateMemoryForMesh(mpfao, dm); CHKERRQ(ierr);
+  // Create the mesh.
+  ierr = TDyMeshCreate(dm, &mpfao->mesh);
+
+  // Read/write connectivity and geometry data if requested.
+  ierr = SaveMeshConnectivityInfo(mpfao, dm); CHKERRQ(ierr);
+  if (mpfao->read_geom_attributes) {
+    ierr = TDyMeshReadGeometry(mpfao->mesh, mpfao->geom_attributes_file); CHKERRQ(ierr);
+  } else {
+    ierr = SaveMeshGeometry(mpfao->mesh); CHKERRQ(ierr);
+  }
+  mpfao->read_geom_attributes = 0;
+  if (mpfao->output_geom_attributes) {
+    ierr = TDyMeshWriteGeometry(mpfao->mesh, mpfao->geom_attributes_file); CHKERRQ(ierr);
+  }
+  mpfao->output_geom_attributes = 0;
+
+  ierr = TDyMeshGetMaxVertexConnectivity(mpfao->mesh, &tdy->ncv, &tdy->nfv);
   ierr = TDyAllocate_RealArray_3D(&mpfao->Trans, tdy->mesh->num_vertices, tdy->nfv, tdy->nfv + tdy->ncv); CHKERRQ(ierr);
   ierr = PetscMalloc(mpfao->mesh->num_faces*sizeof(PetscReal),
                      &(tdy->vel )); CHKERRQ(ierr);
